Retain "unused" bits in conversions to prevent losing information
Byte number 0x06 (the 7th byte) contains four "unused" bits. I do not
understand the significance of these bits.  It seems that they are
sometimes set and sometimes not. If you dump your patches to tape
memory, change these bits and perform "verify" function on the synth
against this changed data, the verification fails. Therefore I want to
save these bits in the conversions (from WAV to CSV and vice versa) just
in case they are somehow significant.

I added one new column to the CSV output. It is called "mystery". The
mystery bits are now used by the audio decoder and encoder.